Rough default radiation shielding values
All armors which had radiation shielding in terms of TechGuns now also have default radiation shielding in NuclearCraft. Values are rougly based on Hazmat Suit parameters of both mods. Sharp drop of values on tier3 (powered) armors is due to ability to add shielding. t2 and lower armors are too shabby to be allowed that.